// apps/admin/src/lib/api/generated/models/deviceLogin.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.3
 */
import type { DeviceLoginStatus } from './deviceLoginStatus';

export interface DeviceLogin {
  expires_at: string;
  login_id: string;
  message?: string | null;
  status?: DeviceLoginStatus;
  user_code?: string | null;
  verification_url?: string | null;
}